Organized taskRequest files into a folder to improve organization
