--- src/main/java/com/playko/projectManagement/application/dto/request/TaskReassignmentRequestDto.java
@@ -1,13 +1,0 @@
-package com.playko.projectManagement.application.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class TaskReassignmentRequestDto {
-    private Long taskId;
-    private Long newUserId;
-}
