+++ src/main/java/com/playko/projectManagement/application/dto/request/task/TaskAssignmentRequestDto.java
@@ -1,0 +1,13 @@
+package com.playko.projectManagement.application.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class TaskAssignmentRequestDto {
+    private Long taskId;
+    private Long userId;
+}
